Improve AppBar/ToolBar logic for displaying the overflow button.
 Change Description:

     closes https://github.com/winjs/winjs/issues/916
closes https://github.com/winjs/winjs/issues/982
closes https://github.com/winjs/winjs/issues/1170
closes https://github.com/winjs/winjs/issues/1210
closes https://github.com/winjs/winjs/issues/1340

This commit adds specialized knowledge to the control about when to show and hide the overflow button and updates the layout pass to accurately reflow commands based on whether or not overflow button has just claimed or released real estate from within the action area.

The logic to show/hide the overflow button was severely out of date, and did not take into account different closedDisplayModes or that commands in the BindingList could be hidden.
Now the overflow button is only hidden when there are no visible commands what so ever, Or when closedDisplayMode === "full" and only the action area has any visible commands.
     Consequently, the private CSS class "win-commandingsurface-empty" is no longer being used to "display: none" the overflow button whenever there are no commands in the binding list
     and has been removed since was un documented and is no longer used. Anyone who may have been using it to determine whether or not the binding list was empty should instead check if
     (winControl.data.length > 0) on the their AppBar or ToolBar.

Improving these main scenarios revealed other implementation failures with the design that have also been addressed, such as including a minimal layout option to make sure we address the overflow button, even if we can't perform a full layout because the control was in a state where it couldn't measure at the time that a layout was requested.

This commit also adds new unit tests and removes or updates many older tests and utilities geared around the appearance of the overflow button based on an out dated model, before the closedDisplayMode property was introduced.

Some redundant tests in  ToolBar and AppBar that were verifying core _Commandingsurface feature areas have been removed. In the future it would be good to keep only ToolBar and AppBar specific tests in the
AppBar and ToolBar test files, and then keep all the common tests for both controls relying on the commandingsurface in the _CommandingSurface test file, and run those test on instances of both ToolBar and AppBar controls.