closes https://github.com/winjs/winjs/issues/1381
This commit changes the role attribute that is being set on "toggle" typed MenuCommands. MenuCommands were using a custom role "menuitemcheckbox" instead of the defacto browser recognized "checkbox" role. Changing the role causes Narrator to correctly announce the updated state of the "aria-checked" attribute on the "toggle" MenuCommand whenever the command is toggle via Narrator