closes https://github.com/winjs/winjs/issues/1400
This commit changes the role attribute that is being set on "toggle" typed AppBarCommands. AppBarCommands were using a custom role "menuitemcheckbox" instead of the defacto browser recognized "checkbox" role. Changing the role causes Narrator to correctly announce the updated state of the "aria-checked" attribute on the "toggle" AppBarCommand whenever the command is toggled via Narrator