Fixing getRelativeLeft
getRelativeLeft was using parentNode, whereas it should be using
offsetParent instead.

We already have a method retrieving the position of an element using
offsetParent, so I've used it here to retrieve the position of the
element and the parent, and then subtracting the latter from the former
gets us the result we want.