Fixing a bug where adding a new PivotItem after removing all PivotItems throws an exception and breaks the first tab
The problem was that setting this._selectedIndex = 0 was being performed
before we had actually added the PivotItem header to the list in
render(), meaning that we were operating on a null header.  Moving that
setting to below render() makes us properly wait until we have a header
before trying to operate on it.