Fixing a bug where a FlipView with >50 items can't be panned from the last item all the way to first item
The problem was that we were only recentering items in the scrollable
area when the next/previous buttons were being pressed, and not when the
user is using touch manipulation.  The fix is simple: we ensure that
we're always recentering items once we've come to a snap point when
using touch manipulation.